pyproject.toml
[build-system]
# Tells pip/build how to build the package
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "WxFrog"
# read version from code
dynamic = ["version"]

authors = [
    { name = "Volker Siepmann", email = "volker.siepmann@gmail.com" },
]
description = "WX-python FROntend kept Generic"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
license = { text = "MIT AND (Apache-2.0 OR BSD-2-Clause)" }
keywords = ["process modelling", "user interface"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "wxpython>=4.2.3",  # needed for several reasons, e.g. DataViewCtrl update
    "pint",
    "pyyaml",
    "Pypubsub"
]

[project.gui-scripts]
wxfrog_hello_world = "wxfrog.examples.hello_world.gui:main"
wxfrog_advanced = "wxfrog.examples.advanced.gui:main"

[project.optional-dependencies]
doc = [
    "Sphinx",
    "docutils",
    "sphinxcontrib-bibtex",
    "sphinx-copybutton",
]
dist = [
    "build",
    "twine",
]
test = [
    "pytest",
]
dev = [
    "pytest",
    "pytest-doctestplus",
    "build",
    "twine",
    "Sphinx",
    "docutils",
    "sphinxcontrib-bibtex",
    "sphinx-copybutton",
]

[tool.setuptools]
# include package data declared via MANIFEST.in or package-data keys
include-package-data = true

[tool.setuptools.package-data]
wxfrog = ["py.typed"]
"*"    = ["src/wxfrog/**/*.yml"]

# dynamic version: same effect as `version = attr: wxfrog._version.version`
[tool.setuptools.dynamic]
version = { attr = "wxfrog._version.version" }